fixing names, organizing in folders
