+++ src/components/molecules/adminManagement/AdminManagementForm.tsx
@@ -1,23 +1,18 @@
 import { observer } from 'mobx-react-lite';
 import React from 'react';
-import DialogWithHeader from './DialogWithHeader';
+import DialogWithHeader from '../DialogWithHeader';
 import { useTranslation } from 'react-i18next';
 import { makeStyles } from '@material-ui/core/styles';
 import { Typography } from 'components/atoms';
 import { useStore } from 'store/storeConfig';
 import RootStore from 'store/root.store';
 import OverlayLoader from 'components/molecules/OverlayLoader';
-import { StyledTableRow, StyledTableCell } from './Table/TableView';
-import { TableContainer, Table, TableBody, TableCell, TableHead, TableRow, Theme } from '@material-ui/core';
+import { StyledTableRow, StyledTableCell } from '../Table/TableView';
+import { TableContainer, Button, Box, Table, TableBody, TableCell, TableHead, Theme } from '@material-ui/core';
 import { blackColor } from 'style';
-import {
-  OrganizationEditList,
-  IeditObjList,
-  EditModeButtons,
-  ViewMode,
-  IProps,
-  labels,
-} from './AdminManagementFormHelpers';
+import { EditModeButtons } from './EditModeButtons';
+import { OrganizationEditList } from './OrganizationEditList';
+import { IeditObjList, IProps, labels } from './AdminManagementFormHelpers';
 
 const useStyles = makeStyles((theme: Theme) => ({
   table: {
@@ -26,7 +21,7 @@ const useStyles = makeStyles((theme: Theme) => ({
     borderCollapse: 'separate',
   },
   tableContainer: {
-    maxHeight: 250,
+    maxHeight: 450,
     paddingBottom: 30,
   },
   saveButton: {
@@ -35,30 +30,29 @@ const useStyles = makeStyles((theme: Theme) => ({
   },
 }));
 
-const AdminManagementForm: React.FC<IProps> = ({ saveEditModeHelper, isShowing, onClose, loading }) => {
+const AdminManagementForm: React.FC<IProps> = ({ saveUserChanges, isShowing, onClose, loading }) => {
   const store: RootStore = useStore();
-  const [editObject, setEditObject] = React.useState<IeditObjList>({});
-
+  const [usersToUpadte, setUsersForUpdate] = React.useState<IeditObjList>({});
   const { t } = useTranslation();
   const classes = useStyles();
 
-  const setEditObjectFunction = (newEditObject: any) => {
-    setEditObject(newEditObject);
+  const setUsersForUpdateFunction = (newUsersForUpdate: IeditObjList) => {
+    setUsersForUpdate(newUsersForUpdate);
   };
 
   const cancelEditMode = (el: any) => {
-    const newObj = { ...editObject };
+    const newObj = { ...usersToUpadte };
     delete newObj[el.email];
-    setEditObject(newObj);
+    setUsersForUpdate(newObj);
   };
   const saveEditMode = (element: any) => {
-    const newObj = saveEditModeHelper(element.email, element.org, { ...editObject });
-    setEditObject(newObj);
+    const newObj = saveUserChanges(element.email, element.org, { ...usersToUpadte });
+    setUsersForUpdate(newObj);
   };
   const changeEditObject = (element: any) => {
-    const newEditObject = { ...editObject };
-    newEditObject[element.email] = { organizationValue: element.name };
-    setEditObject(newEditObject);
+    const newEditObject = { ...usersToUpadte };
+    newEditObject[element.email] = element.name;
+    setUsersForUpdate(newEditObject);
   };
   return (
     <DialogWithHeader
@@ -86,30 +80,34 @@ const AdminManagementForm: React.FC<IProps> = ({ saveEditModeHelper, isShowing, 
           <TableBody>
             {store.usersManagementTableData &&
               store.usersManagementTableData.map((item: any, index: number) => (
-                <TableRow key={index}>
+                <StyledTableRow key={index}>
                   <TableCell align="center">{item.name}</TableCell>
                   <TableCell align="center">{item.email}</TableCell>
                   <TableCell align="center" width="30%">
-                    {editObject[item.email] ? (
+                    {usersToUpadte[item.email] ? (
                       <OrganizationEditList
                         itemOrg={item.org}
                         itemEmail={item.email}
-                        editObject={editObject}
+                        usersToUpadte={usersToUpadte}
                         organizationsList={store.organizationsList}
-                        setEditObjectFunction={setEditObjectFunction}
+                        setUsersForUpdateFunction={setUsersForUpdateFunction}
                       />
                     ) : (
                       <p>{item.org}</p>
                     )}
                   </TableCell>
                   <TableCell align="center" width="40%">
-                    {editObject[item.email] ? (
+                    {usersToUpadte[item.email] ? (
                       <EditModeButtons saveEditMode={saveEditMode} cancelEditMode={cancelEditMode} itemData={item} />
                     ) : (
-                      <ViewMode changeEditObject={changeEditObject} itemData={item} />
+                      <Box>
+                        <Button onClick={() => changeEditObject(item)} variant="contained">
+                          {t('usersManagement.edit')}
+                        </Button>
+                      </Box>
                     )}
                   </TableCell>
-                </TableRow>
+                </StyledTableRow>
               ))}
           </TableBody>
         </Table>
